+++ src/com/tutorial/game/FastEnemy.java
@@ -14,7 +14,7 @@ public class FastEnemy extends GameObject {
     }
 
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 16, 16);
+        return new Rectangle((int) x, (int) y, 16, 16);
     }
 
     public void tick() {
@@ -24,11 +24,11 @@ public class FastEnemy extends GameObject {
         if (y <= 0 || y >= Game.HEIGHT - 32) velY *= -1;
         if (x <= 0 || x >= Game.WIDTH - 16) velX *= -1;
 
-        handler.addObject((new Trail(x, y, ID.Trail, Color.CYAN, 16, 16, 0.05f, handler)));
+        handler.addObject((new Trail((int) x, (int) y, ID.Trail, Color.CYAN, 16, 16, 0.05f, handler)));
     }
 
     public void render(Graphics g) {
         g.setColor(Color.CYAN);
-        g.fillRect(x, y, 16, 16);
+        g.fillRect((int) x, (int) y, 16, 16);
     }
 }
